refactor: move vision schemas to shared schemas package
